[CBRD-24378] Fix the restoredb not to go beyond the given time (-d) (#3804)
http://jira.cubrid.org/browse/CBRD-24378

- Purpose
To prevent the restoredb from doing recovery beyond the given time with the -d option.

- Implementation
A new log record for the sysop commit is added. It includes the donetime like the LOG_COMMIT. This time information is used for the time-specific restore.
For backward compatibility, the old log record type still remains.
The new log record type uses the name of the old log record type name: LOG_COMMIT_WITH_POSTPONE.
The name of the old record type is changed to LOG_COMMIT_WITH_POSTPONE_OBSOLETE.
The macro number that the new log record type uses was for LOG_WILL_COMMIT, which had been obsolete for a long. I believe no one has a log including this type.