[CBRD-24461] Fix the core that occurs when finding user information in cdc (#3775)
http://jira.cubrid.org/browse/CBRD-24461

Background

supplemental log for transaction user is logged at the begin of the transaction and at the end of the transaction
In CDC, when transaction user log is not found at the begin of the transaction, it search the transaction user info which is logged at the end of the transaction

Purpose

When a supplemental log related to serial is logged, the log which contains transaction user info may be omitted at the begin of the transaction. (serial for auto_increment case)
The core occurs while traversing the log records to find the tran user information that was recorded at the end of the transaction. This is because the page at the time it was copied to the local buffer does not have the latest record at this time.
When referring to the past log, the transaction user information recorded at the commit time is used only when the transaction start point is truncated.
So, if requested transaction is not committed, then cdc_find_user() returns fail