[CBRD-24473] Index scan cannot be used when a stored function is used in a where condition. (#3811)
http://jira.cubrid.org/browse/CBRD-24473

I fixed it so that the method is treated as a pseudo constant in qo_analyze_terms().

'col1 = const' <== it is indexable.

There is one more problem. Index scan on correlated columns is not possible due to query rewrite in meth_translate() (check test cases for CBRD-24473.txt). I hope this will be fixed in the future.