[LETS-548] Page Server collects and aggregates MVCC information received from Passive Transaction Servers (#3837)
http://jira.cubrid.org/browse/LETS-548

Page Server must collect and aggregate MVCC-related statuses received from all connected (alive) Passive Transaction Servers in order to be able to return the status to a request coming from the Active Transaction Server.

No error handling regarding PTS process (crashing, being closed) is done in this story. This is for later.

There is a dangerous window between when a new PTS is connected and when the first message containing the oldest mvccid from it is received. If the ATS does vacuum without considering it, data seen by some transaction on the newly connected PTS can be cleaned up. See the scenario: 
 (1) a new PTS is connected and it sends the first oldest mvccid.
 (2) the PTS executes a transaction referring to the mvccid
 (3) the PS hasn't received the message and ATS vacuums it.
 (4) the PS receives the message.
To address this, we create the oldest mvccid entry when PTS connects to PS and set it to MVCCID_ALL_VISIBLE, which means "do not vacuum". If the ATS sees this, it gives up the vacuum and retries later.