[CBRD-24506] Changing the way characters are read from DBMS (#3880)
http://jira.cubrid.org/browse/CBRD-24506

Oracle/MySQL has a character set of the server and a character set of the client, respectively.
There is a risk that the server's character set and the client's character set do not match, or an error may occur whenever the character set is changed. To prevent such an error, it is necessary to change it to read in unicode using a unicode driver.