[LETS-564] Store channel id when connection between PS and TS is made successfully (#3883)
http://jira.cubrid.org/browse/LETS-564

Purpose

PS need to find and remove the oldest active mvcc id of a specific PTS using the channel ID, but when PS try to get the channel ID while the connection is disconnected, an invalid channel ID is created. This is because the socket value is already initialized to -1.

Therefore, rather than configuring the channel ID every time PS tries to get the channel ID, it is better to use the stored channel id which is stored when it is connected successfully.