[LETS-560] [Regression] Broken PTS-PS page synchronization mechanism causes crash in PS (#3888)
http://jira.cubrid.org/browse/CBRD-560

Revert LETS-458, namely #3819.

The page allocation contains several works and it’s done by setting the page type, however, they generate separate log records. The RVBT_GET_NEWPAGE on the issue is one example. The page is actually already allocated (volume sector table or file sector table are already modified), but when trying to set the page type, it is failed to fix the page because the page has no page type yet. It probably happened in all the cases where allocating a new page is replicated on PTS.
We'll come up with a solution considering this case.