[LETS-550] Load vacuum data first to prevent a potential bug (#3904)
http://jira.cubrid.org/browse/LETS-550

Small fixes:
- The vacuum data is loaded on the master thread loop, not when the vacuum is initialized.
There is a low chance the data is not loaded, and in turn, it dumps core when it is shutdown if the vacuum master thread job returns continuously without loading the data. It can happen if connection to PS fails continuously. So, I fix the data to be loaded before any job. It would barely happen, but what is correct is correct.
- Remove an assert that says get_active_tran_server_ptr ()->get_oldest_active_mvccid_from_page_server () always succeeds. The connection with PS may not work for a while for some reason. However, from the vacuum's point of view, the connection's stability doesn't need to be guaranteed. In that case, it can just restart later with the expectation it will get stable.