[LETS-595] overflow heap page atomic sequence (#3954)
http://jira.cubrid.org/browse/LETS-595

Passive Transaction Server atomic replication sequence for page overflow allocation scenarios.

Avoid purging an atomic sequence by an `LOG_SYSOP_END` (with `LOG_SYSOP_END_LOGICAL_UNDO`) when that sequence was not started by a `LOG_SYSOP_ATOMIC_START` (see log sequence synopsis in the comment for `class atomic_replication_helper`).

Also drop the condition regarding `lastparent_lsa` to fit both the cases when it has a valid value and when it has a null value. This is needed to accommodate the situation when `LOG_SYSOP_ATOMIC_START` log record that starts the atomic sequence coincides with the start of the transaction (is the very first log record of a newly occurring transaction in the log).