[LETS-458] Revert the reverted change for LETS-458 (#3996)
http://jira.cubrid.org/browse/LETS-458

It is actually to revert the reverted 6d6e2cb. The side effect of that commit has been fixed by other commits.

The below is from 6d6e2cb:

The cause of the hang: When a page is requested with a target lsa by PTS, the target lsa set to the m_redo_lsa, which is the next one to be applied. PS waits for the replication to go past the target LSA and returns the page. If the last redo record is applied, the target lsa points to after the last log record and the PS can't go past it because no log record is generated without ATS.

Set the target lsa by PTS to the last applied LSA, not the one that will be applied.