[CBRD-24252] Removes the table of the primary key from the join with the table with the foreign key referencing it. (#3838)
http://jira.cubrid.org/browse/CBRD-24252

A table with a primary key is removed from a join with a table with a foreign key that references the primary key. There must be no references to the table to be removed other than join predicates. This is because the relationship between a primary key and a foreign key replaces a data filter of a join. If a table with a primary key is removed, a predicate for IS NOT NULL may be added, as it cannot filter on NULL.