[LETS-639] Refactor async_disconnect_handler to use it in tran_server too (#4132)
http://jira.cubrid.org/browse/LETS-639

We need to add the "disconnection request" from PS to TS. In the request handler, it will try to delete connection_handler object, but it ends up kind of deadlock. The request handler is called by the internal request receiver thread in connection_handler. If it tries to delete itself, it tries to join itself in the dtor (~connection_handler). To address this issue, we can pass the deletion job to another thread. At first, I come up with using a detached thread, but the async_disconnect_handler has been introduced seemingly for this purpose in #3407 and a more robust way. So, I make it a template class to be used in not only page_server but also tran_server.

- No functional change
- Make the async_disconnect_handler a template class to be used in common for tran_server and page_server.