[CBRD-24633] Delete owner information from schema, index, trigger, and object files of general users. (#4136)
http://jira.cubrid.org/browse/CBRD-24633

If there is owner information in the schema file, an error may occur when moving data from user_a to user_b. In this case, it is necessary to delete the owner information or change the onwer information in the schema file. In order to remove this inconvenience, it is necessary to delete the owner information.