[CBRD-24662] Add a LoggingThread to access javasp's log file by only one thread (#4137)
http://jira.cubrid.org/browse/CBRD-24662

Java FileHandler creates a new file with an increasing number when there exist conflicts to a file. Because of this, If an exception occurs very quickly by many threads in the Java SP server, Many log files are generated by the FileHandler's behavior.
To avoid this behavior, This PR implements the LoggingThread to access a log file with a single thread. Other threads will put their exception into LinkedBlockingQueue. And the LoggingThread will take the log string of exception without conflict.