[LETS-659] response broker blocking terminate (#4161)
http://jira.cubrid.org/browse/LETS-659

Response broker termination function is extended into a blocking call - `notify_terminate_and_wait`.

In order to allow for this extension to work correctly it was needed to keep track of en-route requests to the peer server. Up until now, an en-route request was not marked in any way. Extended the broker with a `register_request` call which actually add and entry in the appropriate broker bucket before actually dispatching the roundtrip request, thus, allowing the terminating logic to avoid race conditions.
Adapted and extended the unit test as a result of this behavioral change.

The sync client server is modified to use the blocking call.

Currently this function is only explicitly invoked on the Page Server side. With subsequent refactorings, the functionality will be used on Transaction Servers as well.

Other:
- reworded some comments