[CBRD-24708] Fix type binding of parameterized PL/CSQL variable in Static SQL (#4239)
http://jira.cubrid.org/browse/CBRD-24708

Host variables, not character types, have a null expected domain, so the type binding of parameterized PL/CSQL variable name is not properly resolved.

If the lhs of the expression which has a rhs host variable is not string type, The expected domain is set as tp_null (null domain). In the db_push_values (setting host variable value when executing the query), If the expected domain is null, It just tries to coerce the value at the time. So I've changed the routine to bind the host variable's type in PL/CSQL compiler module.