[LETS-666] add extra perfmon metric wether slotted pages need compacting (#4250)
http://jira.cubrid.org/browse/LETS-666

Add extra information to the existing `PERF_SERVED_COMPR_` perf extended statistics for whether the the slotted pages need also compacting.

Implementation:
- added enum value `PERF_SERVED_COMPR_SLOTTED_NEEDS_COMPACT`
- when a page is served from Page Server to a Transaction Server, check whether the page is slotted, and - if it is - whether it needs to be compacted and - if specified via the `PERFMON_ACTIVATION_FLAG_SERVED_COMPR_PAGE_TYPE` - also log this information
- extended perfmon functions to also add this information

Context:
In the monolithic server, it is the job off the vacuum daemons to ensure that the slotted pages remain in an optimal state. The reason for this might be that, chronologically, the vacuum is the last entity to touch a modified page after all the other (the modifying transaction, potential older concurrent transactions that might also need older values, ...).
In the scalability architecture, the vacuum threads compacting the page - although currently still done - it is useless, as the Active Transaction Server pages, compacted or not, will just be evicted from the page buffer after a certain time.
In the scalability architecture, in order to have the on-permanent-storage data pages also in an optimal state, it is needed that the replication threads on Page Server perform this action because these threads are the only ones that actually modify the data pages.