[LETS-672] Change the main connection when the previous one is disconnected. (#4285)
http://jira.cubrid.org/browse/LETS-672

One of the connections (PS) is regarded as the main connection, from which a TS elicits data/log pages. When the main connection is disconnected, another one must take over the role. In the case of PTS, log prior nodes are sent through the connection, so it's quite complicated. In this issue, we only address the ATS case.

* Design choice *
- Open the main connection concept to upper layers (who use tran_server) and throw the main connection changing err msg to them to deal with it in their own way.
- Hide the main connection concept to upper layers and make send_receive communicate with the main connection in one way or another.

I've chosen option (2) because the upper layers don't need to know the details and there are no other meaningful reactions in the case that the main connection is being changed. Welcome to any argument.