[LETS-680] Handle the no-ps-conn error case in the callers (#4287)
http://jira.cubrid.org/browse/LETS-680

When no PS connection is available, send_receive fails, but it isn't necessary and is bothersome in some cases. In this PR, the error case is got out of send_receive and handled in the caller's own way. The ER_CONN_NO_PAGE_SERVER_AVAILABLE is added.

Following the existing error handling like other errors (no exception handling in the callers):
- SEND_LOG_BOOT_INFO_FETCH: logpb_fatal_error
-GET_OLDEST_ACTIVE_MVCCID: just print an error msg and retry later 
- SEND_STOP_LOG_PRIOR_DISPATCH: ignore
- GET_BOOT_INFO: fails to boot.

assert_release:
- SEND_LOG_PAGE_FETCH: assert_release
- SEND_DATA_PAGE_FETCH: assert_release