[CBRD-24785] [PLCSQL] Fix the LIMIT clause in Static SQL does not work. (#4333)
http://jira.cubrid.org/browse/CBRD-24785

In the parser, the LIMIT clause in static SQL is converted to a numbering expression in a form that cannot be compiled.
- Apply PT_CONVERT_RANGE for a rewritten query to be able to compile.
- During query parsing, a LIMIT clause is converted to a numbering expression node (inst_num(), groupby_num(), orderby_num())  which is added to the search condition clauses (WHERE, ORDERBY_FOR..) without removing the original LIMIT clause node. This is done because the LIMIT clause node is used in XASL generation. However, when you print the entire parser node, the LIMIT node is hidden because the query parser treats it as already rewritten.
In the next phase, the query rewriter performs auto-parameterization on the LIMIT clause and the numbering expression node. This creates two host variables that correspond to the LIMIT and numbering expression nodes. However, since the LIMIT clause node is hidden, its host variable loses its pointer.
In static sql parsing mode, the limit doesn't have to auto parameterization because it will not be executed immediately. Therefore static sql parsing mode skips qo_do_auto_parameterize_limit_clause() in query_rewrite.c.