[CBRD-24870] Increase the number of sampling pages when collecting btree statistics (#4448)
http://jira.cubrid.org/browse/CBRD-24870

Change the number related to sampling statistics for b-tree.
STATS_SAMPLING_THRESHOLD : 50 to 1000. If this value is not exceeded, the entire page is full-scanned.
STATS_SAMPLING_LEAFS_MAX : 7 to 500. The number of samples extracted when collecting statistics on sampling.
Since the number of leaf pages in the b-tree is smaller than the number of pages in the heap, the number of b-tree samples is set to 500. The number of sampling pages in the heap is 1000. These values will be adjusted later in optimizer enhancement.

Change sample weights from int to double for accurate calculation.
Delete routine that do not update statistics on the same number of pages to prevent the problem that statistics are not updated.
Delete routine for acceptance probability. Since the number of sampling pages is increased, it becomes unnecessary.