[LETS-697] Handle the abnormal disconnection of a page server. (#4455)
http://jira.cubrid.org/browse/LETS-697

An error during communication is not spotted immediately. It's noticed when a request is pushed to the connection. The disconnection can be found when

(A) push_request() : it's reported through the error handler given when setting the connection (connection_handler::set_connection).
send_receive() :
(B) while pushing a request: it's reported through the error handler given in the request_sync_client_server::send_recv
(C) while getting its response: it's never reported now.

When the connection is disconnected, the push() fails, its internal socket gets invalid (channel:send), and throws the error. The error code, in this case, is ERROR_ON_WRITE, but I set it to CONNECTION_CLOSED when the connection gets invalid to let the outside know that it's not just a write error, but the socket has been closed. And in each case:

(A): If the error code is CONNECTION_CLOSED, clean up the connection_handler.
(B): The current error handler set the error code and make it's returned to get_response(). So, I leave it as it is and deal with the case it's returned in connection_handler::send_receive. It's because request_sync_client_server::send_recv doesn't know connection_handler.
(C): If push() succeeded and the connection fails while waiting for its response, there is no way for the waiter to know it. The subsequent request will fail, clean the connection_handler, and, in turn, let the waiter know the connection_handler got wrong so it's closed. In this case, the default error code, ERROR_ON_WRITE is returned. This is a walkaround and this should be addressed with a timeout for send_receive or something in a separate issue.