[LETS-742] streamline scalability communication protocol (#4408)
http://jira.cubrid.org/browse/LETS-742

The connection support layer (`css_*` functions) already send the length and the data (as an application level protocol).
In the network code introduced by LETS (`communication_channel`, `request_client_server`), the sending of length was done ome more time.
This translated in 4 calls to underlying `send`/`writev` functions:
1. send/recv "length of length"
2. send/recv actual length as data
3. send/recv "length of data"
4. send/recv actual data

This was done in the beginning of development for speed/simplicity of development rather than performance of the data transfer.
It is easy to see that 1. and 2. are superfluous.
The benefit of having 1. and 2. is that they are both messages of fixed/known length and they allowed the receiving side to know exactly how much data it would receive in step 4.
Removing them has the downside that the receiving side's application layer never knows in advance how much data it receives in step 4.
But also, removing them, has the benefit of uncluttering the dialog between peer servers.
To remove threm, the connection support layer has been extended to allow a two step receive of unknown side data (see functions `css_net_recv_allow_truncated` and `css_net_recv_remainder` - patch 8253662fb) such as to allow proper consumption of data in step 4.
In practice, in the receiving side's application layer - with the reuse of a memory buffer (in this case `cubmem::extensible_buffer`) the calls to the second function will be very rare (if ever).

Implementation:
- `channel::recv_allow_truncated`: calls into `css_net_recv_allow_truncated`
  - if all data fits into supplied buffer it will return a `NO_ERRORS` code and output a 0 `remaining_len` (same as `channel::recv` does)
  - if not all data fits into supplied buffer, it will return a `RECORD_TRUNCATED` code and a non-zero value for `remaining_len`
- if the application layer receives `RECORD_TRUNCATED` it is supposed to follow-up with a call to `channel::recv_remainder` and a properly sized buffer as to consume all remaining data as dictated by the previous call to `recv_allow_truncated`
- `request_server::receive_request_buffer` is the function that actually implements this simplification - the `recv_allow_truncated` is followed up by a call to `recv_remainder` when/if needed
  - the signature of `receive_request_buffer` is unchanged so nothing has changed for the upper layers
- adapted the relevant unit test to also exploit/test the newly added mechanism

Other:
- the signature difference for function `send_client_request` seems to be a merge error from a previous patch (31803efd1)