[LETS-746] Add the request to send the target PS and catch-up LSA (#4474)
http://jira.cubrid.org/browse/LETS-746

* catchup_lsa: unsent_lsa in LOG_PRIOR_LSA_INFO.
It is the prior_lsa when the prior list is sent, from which new prior nodes will be appended and sent to PSes after sending the current ones. This is the LSA of the first prior node which hasn't been sent yet.
It's updated when sending prior lists so that the new PS can know the first LSA it will receive when adding its sink_hook.
- It is taken when prior lists are removed to send in logpb_prior_lsa_append_all_list.
- It is set in log_Gl.m_prior_sender() when sending.
- It is retrieved and used as the catchup_lsa while adding sink_hook in on_connecting.

* The SEND_START_CATCH_UP request
- Send (target ps info { hostname, port }, catchup_lsa)
- It's sent after adding the sink_hook in on_connecting.
- dummy info is sent while the ATS is booting. See the Note.

* Note
We don't deal with sending the right information in the booting-up case. It will be receive_start_catch_up: hostname: N/A, port : -1, LSA (-1|-1).