[LETS-782] Create connection handlers to catch up between PSes (#4611)
http://jira.cubrid.org/browse/LETS-782

When the connection request from PS to PS is received, a connection_handler for the catchup should be created. It’s temporary, which means it’s only alive during the catch-up.

Creating a separate connection_handler rather than extending the one for the tran_server, we can keep it simple only with what's necessary.
- We don't need the asynchronous responder because, during the catch-up, the follower requests sequentially.
- We don't need the prior_sender, of course.
- The disconnection protocol is much simpler. Some handshake is needed between PS and TS because a node has to make sure no more messages will be sent, or the other side regard it as an abnormal disconnection. However, in the case of the catchup_connection_handler, only the follower makes requests and the followee responds to them, the follower can decide to disconnect no matter when.
- Wrapping up, in my opinion, we can keep it simple and safe with dedicated connection_handlers for the catch-up process which only has what's needed.

Splitting the catchup_connection_handler into follower_connection_handler and followee_connection_handler, it's safer and clearer. Some of features are only for followee and some are for follower. If we bundle them up, we have to keep in mind that which functions are for which one. It's vulnerable to human error. For example, the way to disconnect will be different, and only follower makes requests and followee just responds to them.