[LETS-783] Have waiters on send_receive notice the abnormal disconnection (#4598)
http://jira.cubrid.org/browse/LETS-783

Waiters who are blocked on send_receive() can’t wake up even when their connection is broken down. Instead, they are notified when another thread tries to push a request, notice it’s broken, and close it. Shortly, receiving can’t be notified of its disconnection while sending can be.

This can be serious when no following request is pushed. The blocked operations by send_receive can't go forward. Usually, ATS sends log records and PTS sends MVCC information, but it may not happen when there is no activity on them. Moreover, the inter-ps connection for the catch-up does not have this kind of request, so it can be blocked when the connection is broken down.

In this issue, we will provide a way to let the waiter to know about the abnormal disconnection. Currently, when the request receiver that process replies just exits its handler loop when it runs into an error during receiving continuously pushed requests.

To deal with this, we are going to push an error_handler called when something gets wrong while reading a request (respond) from a socket.