[CBRD-24974] The problem is that error occurs when loading unloaded schema file if both DBA and user have tables and views with same name (#4624)
http://jira.cubrid.org/browse/CBRD-24974

The `unloaddb` utility outputs CREATE statements for tables and views in the schema file, where the original user becomes the owner as soon as created. DBA does not temporarily own tables and views, and the `change_owner ()` method is not called. The `emit_class_owner()` function has been removed as it is deprecated.