[CBRD-25001] When executing SQL in the gateway, there are cases where the affected row is unknown. (#4698)
http://jira.cubrid.org/browse/CBRD-25001

when using the ODBC Driver, there are cases where the number of rows affected during query execution is unknown. At this time, ODBC returns the number of rows as -1.
If CUBRID returns -1 as a result code to the client, an error occurs as CCI repeatedly calls connection, prepare, and execute.
If the affected number is -1, an error is processed and the user is notified that the affected number is unknown.
Error message : Number of rows affected is unknown.