[LETS-751] Notify to the ATS that the catch-up is completed (#4722)
http://jira.cubrid.org/browse/LETS-751

The ATS doesn't know if a PS is catching up or not. It just regards them as lagging behind. The saved_lsa is sent to the ATS when a log page is flushed and it’s also the case during the catch-up. When the saved_lsa from a PS goes past the catchup_lsa, it means that the PS is done catching up and can attend the quorum. However, it is helpful to notify when the job is done from time to time. An example is the main connection change. A PS is not allowed to be the main connection during the catch-up.