[CBRD-25005] Semantic API returns error when a field lookup expression is used in Static SQL (#4729)
http://jira.cubrid.org/browse/CBRD-25005

To handle cursor variables, I've changed PT_PARAMETER note to PT_HOST_VARIABLE node in PL/CSQL compile mode. And the name of a cursor variable is stored in the host variable's label. It is sent to the PL/CSQL compiler to bind variable's name from PL/CSQL's symbol table.

---------
Co-authored-by: Na, Hyunik <hyunikn@gmail.com>