[LETS-681] The checkpoint without the quorum confirmation (#4294)
http://jira.cubrid.org/browse/LETS-681

The checkpoint on ATS requires quorum confirmation, which is needed to synchronize the checkpoint information on ATS (m eta file) and the log records on PS. This can cause a hang when an ATS stops with PSes less than quorum because the checkpoint daemon can't get its job done. This quorum requirement for the checkpoint is going to be removed when Recovery references the checkpoint information on PS. So, we're getting rid of this restriction in advance to avoid the hang.