[CBRD-25049] Core dumped in mvcc_is_mvcc_disabled_class at src/transaction/mvcc.c:618 (#4757)
http://jira.cubrid.org/browse/CBRD-25049

This issue is due to the changes made by the CBRD-25037 issue. The error occurred because the handling of the case where class_oid, used as an argument in mvcc_is_mvcc_disabled_class(), is NULL before executing mvcc_is_mvcc_disabled_class() was not handled. Therefore, modifications have been made to handle this case.