[CBRD-25048] Core dumped in pgbuf_replace_watcher_debug at src/storage/page_buffer.c:12574 (#4756)
http://jira.cubrid.org/browse/CBRD-25048

This issue arises due to the modifications induced by the CBRD-25037 issue. The reason for the error is that the processing within the heap_next_internal () differs between when the select_record_info hint is used and when it is not used. Therefore, when the select_record_info hint is used, modifications have been made to transfer the value to the local page_watcher variable as before the change, and then execute heap_get_record_info ().