[CBRD-25041] Fix segfault when a arg of function is undefined parameter. (#4741)
http://jira.cubrid.org/browse/CBRD-25041

If :a is not defined, "Query yields no result, so variable 'a' is not set." should be returned. But the undefined parameter (:a, PT_NAME node with PT_PARAMETER meta class) was not propely handled before checking the type of paramter. So even though the error should occurs, it continues and when checking the argument {:a}, segfault occurs at assertion (arg's type is not checked well, PT_TYPE_NONE is specified).