[LETS-873] Temporary solution for PTS read transactions to wait when accessing a DROPping table (#4811)
http://jira.cubrid.org/browse/LETS-873

Purpose
locator_is_exist_class_name_entry() is called during PTS read transactions checking if a table exists in both the cache and heap. The assertion failure occurs in ATS when the cache state is LC_CLASSNAME_EXIST but the table is not in the heap. In the context of PTS, since the cache state doesn't guarantee the presence of a heap record, this assertion is unnecessary. However, this is a temporary workaround, and in the future, when implementing logic to update the cache state, this change will be reverted