[CBRD-25104] [Regression] query cost is changed in query plan (#4824)
http://jira.cubrid.org/browse/CBRD-25104

There is an issue where the total number of rows in the partition table are added per partition. the total number of rows is used without adding.
There is an issue where the selectivity is set to 1 when the varchar is over 4000. it is fixed to be set to 1 only when NDV is 0.