[LETS-806] Fix the core when the RVPGBUF_DEALLOC log record is replicated to PTS (#4795)
http://jira.cubrid.org/browse/LETS-806

Purpose

When applying RVPGBUF_DEALLOC in the PTS replicator, the redo function already invalidates and unfixes the corresponding page.
However, because the replicator is unaware of whether the page has been unfixed, it attempts to unfix it again. The goal is to resolve the crash occurring during this process.
If the page has already been unfixed, it is necessary to invalidate rcv.pgptr used by the replicator.